+++ json-rest-mongodb/src/main/java/com/spring/project/json/service/CabinServiceImpl.java
@@ -1,5 +1,6 @@
 package com.spring.project.json.service;
 
+import com.spring.project.json.handler.CabinNotFoundException;
 import com.spring.project.json.model.Cabin;
 import com.spring.project.json.repository.CabinRepository;
 import jakarta.persistence.NoResultException;
@@ -23,7 +24,7 @@ public class CabinServiceImpl implements CabinService{
 
     @Override
     public Optional<Cabin> findByCabinId(String id) throws NoResultException {
-        return cabinRepository.findById(id);
+        return Optional.ofNullable(cabinRepository.findById(id).orElseThrow(() -> new CabinNotFoundException(id)));
     }
 
     @Override
@@ -44,11 +45,14 @@ public class CabinServiceImpl implements CabinService{
 
                     return cabinRepository.save(existingCabin);
                 })
-                .orElseThrow(() -> new NoResultException("Cabin not found with id: " + id));
+                .orElseThrow(() -> new CabinNotFoundException(id));
     }
 
     @Override
     public void delete(String id) {
+        if (!cabinRepository.existsById(id)) {
+            throw new CabinNotFoundException(id);
+        }
         cabinRepository.deleteById(id);
     }
 }
